+++ processing/src/Parser/Parser.h
@@ -8,13 +8,13 @@ class Parser {
 	public:
 		Parser(const char* file);
 		~Parser();
-		double*** getData();
+		int*** getData();
 		unsigned int getColumnSize();
 		unsigned int getRowSize();
 
 	private:
 		const char* file;
-		double*** data;
+		int*** data;
 		unsigned int columnSize;
 		unsigned int rowSize;
 		void parseDataSize();
